[refactor] moving InsertPostInitMethodToModuleSubClasses to utils.
